requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-docx==0.8.11
PyPDF2==3.0.1
langchain==0.1.0
tiktoken==0.5.1
python-dotenv==1.0.0
openai==1.12.0
pymupdf==1.23.0
